vandelay_migration/pyproject.toml
[project]
name = "vandelay-migration"
version = "1.0.0"
description = "Vandelay Migration Agent - VMware to BareMetal OpenShift Migration Assistant"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=0.1.0",
    "litellm>=1.0.0",
    "neo4j>=5.0.0",
    "httpx>=0.24.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.20.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vandelay_migration"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]
